Restructured table components into directories & renamed
